Moved each roles to a separate folder
